实现了ESP8266的UDPDP_SOCKET收发功能

+++ Demo/include/userFlash.h
@@ -7,14 +7,15 @@
 #define __USER_FLASH_H__
 
 #ifdef __cplusplus
-extern "C" {
+extern "C"
+{
 #endif
 
 #include "userFlashMap.h"
 
-extern int Fun_Flash_Erase(unsigned int sector_num);
-extern int Fun_Flash_WR(int offset,unsigned int *pBuf,unsigned int nBytes);
-extern int Fun_Flash_RD(int offset,unsigned int *pBuf,unsigned int nBytes);
+    extern int Fun_Flash_Erase(unsigned short sector_num);
+    extern int Fun_Flash_WR(int offset, unsigned int *pBuf, unsigned int nBytes);
+    extern int Fun_Flash_RD(int offset, unsigned int *pBuf, unsigned int nBytes);
 
 #ifdef __cplusplus
 }
